refac(extract): move selection to separate module
